Creating an example project
