+++ StepIndicator.js
@@ -1,200 +1,184 @@
 
 import React, { Component,PropTypes } from 'react';
-import { View,Text,StyleSheet, Dimensions, Animated, Easing } from 'react-native';
-const STEP_COUNT = 5;
-const STEP_INDICATOR_SIZE = 30;
-const SELECTED_STEP_INDICATOR_SIZE_DELTA = 10;
-const SEPARATOR_STROKE_WIDTH = 10;
-const STEP_INDICATOR_COLOR_CURRENT = '#ffffff';
-const STEP_INDICATOR_COLOR_FINISHED = '#4aae4f';
-const STEP_INDICATOR_COLOR_UNFINISHED = '#a4d4a5';
-const SEPARATOR_COLOR_FINISHED = '#4aae4f';
-const SEPARATOR_COLOR_UNFINISHED = '#a4d4a5';
-const STEP_INDICATOR_LABEL_SIZE = 15;
-const STEP_INDICATOR_LABEL_COLOR_CURRENT = '#000000';
-const STEP_INDICATOR_LABEL_COLOR_FINISHED = '#ffffff';
-const STEP_INDICATOR_LABEL_COLOR_UNFINISHED = 'rgba(255,255,255,0.5)';
+import { View,Text,StyleSheet } from 'react-native';
+
+var customStyles = {
+  stepIndicatorSize: 30,
+  currentStepIndicatorSize:40,
+  separatorStrokeWidth: 3,
+  currentStepStrokeWidth: 5,
+  separatorFinishedColor: '#4aae4f',
+  separatorUnFinishedColor: '#a4d4a5',
+  stepIndicatorFinishedColor: '#4aae4f',
+  stepIndicatorUnFinishedColor: '#a4d4a5',
+  stepIndicatorCurrentColor: '#ffffff',
+  stepIndicatorLabelFontSize: 15,
+  currentStepIndicatorLabelFontSize: 15,
+  stepIndicatorLabelCurrentColor: '#000000',
+  stepIndicatorLabelFinishedColor: '#ffffff',
+  stepIndicatorLabelUnFinishedColor: 'rgba(255,255,255,0.5)'
+}
 
 export default class StepIndicator extends Component {
 
   constructor(props) {
     super(props);
-    this.animatedValue = new Animated.Value(0);
     this.state = {
       width:0
     }
+    customStyles = Object.assign(customStyles, props.customStyles);
   }
-
-  componentWillReceiveProps(newProps) {
-    this.animateStep();
-    console.log("ANIMATED");
-  }
-
 
   render() {
+    const { labels } = this.props;
     return (
-      <View onLayout={(event) => this.measureView(event)} style={styles.container}>
+      <View onLayout={(event) => this.setState({width: event.nativeEvent.layout.width})} style={defaultStyle.container}>
         {this.renderStepIndicator()}
+        {labels && this.renderStepLabels()}
       </View>
     );
-  }
-
-  measureView(event) {
-    var {width} = Dimensions.get('window');
-    this.setState({
-      width: event.nativeEvent.layout.width,
-    })
   }
 
   renderStepIndicator = () => {
     let steps = [];
-    for(let position = 0 ; position < STEP_COUNT ; position++) {
-
-      steps.push(<View key={position} style={styles.stepContainer}>
-        {this.renderStep(position)}
-        {position !== STEP_COUNT - 1 && this.renderStrip(position)}
-      </View>)
-    }
-    return(
-      <View style={styles.stepIndicatorContainer}>
-        {steps}
-      </View>
-    )
-  }
-
-  renderStep = (position) => {
-    const { currentPosition } = this.props;
-    let stepStyle;
-    let stepLabelStyle;
-    let selectedStepStyle;
-    const stepSize = this.animatedValue.interpolate({
-      inputRange: [0, 1],
-      outputRange: [STEP_INDICATOR_SIZE, STEP_INDICATOR_SIZE + SELECTED_STEP_INDICATOR_SIZE_DELTA]
-    });
-    const stepBorderRadius = this.animatedValue.interpolate({
-      inputRange: [0, 1],
-      outputRange: [STEP_INDICATOR_SIZE/2, (STEP_INDICATOR_SIZE + SELECTED_STEP_INDICATOR_SIZE_DELTA)/2]
-    });
-    if(position === currentPosition) {
-      stepStyle = styles.stepCurrent;
-      stepLabelStyle = styles.stepLabelCurrent;
-      //console.log("STYLE " , selectedStepStyle);
-
-    }
-    else if(position < currentPosition){
-      stepStyle = styles.stepFinished;
-      stepLabelStyle = styles.stepLabelFinished;
-      //selectedStepStyle={height:STEP_INDICATOR_SIZE , width:STEP_INDICATOR_SIZE, borderRadius:STEP_INDICATOR_SIZE/2};
-    }
-    else {
-      stepStyle = styles.stepUnFinished;
-      stepLabelStyle = styles.stepLabelUnFinished;
-      //selectedStepStyle={height:STEP_INDICATOR_SIZE , width:STEP_INDICATOR_SIZE, borderRadius:STEP_INDICATOR_SIZE/2}
-    }
-    selectedStepStyle = {height:stepSize , width:stepSize, borderRadius:stepBorderRadius}
-
-
-    return (
-      <Animated.View style={[styles.step , stepStyle ]}>
-        <Text style={[styles.stepLabel , stepLabelStyle]}>{ position + 1 }</Text>
-      </Animated.View>
-    );
-  }
-
-  animateStep = () => {
-    console.log("ANIMATE STEP");
-    this.animatedValue.setValue(0);
-    Animated.timing(
-      this.animatedValue,
-      {
-        toValue: 1,
-        duration: 100,
-        easing: Easing.exp
+    const { labels, stepCount } = this.props;
+    for(let position = 0 ; position < stepCount ; position++) {
+      steps.push(
+        <View key={position} style={defaultStyle.stepContainer}>
+          {this.renderStep(position)}
+        </View>)
       }
-    ).start()
+      return(
+        <View style={defaultStyle.stepIndicatorContainer}>
+          {steps}
+        </View>
+      )
+    }
+
+    renderStepLabels = () => {
+      const { labels } = this.props;
+      var labelViews = labels.map((label,index) => {
+        return (
+          <Text key={index} style={defaultStyle.stepLabel}>
+            {label}
+          </Text>
+        )
+      });
+
+      return(
+        <View style={defaultStyle.stepLabelsContainer}>
+          {labelViews}
+        </View>
+      )
+    }
+
+    renderStep = (position) => {
+      const { currentPosition, stepCount } = this.props;
+      let stepStyle;
+      let indicatorLabelStyle;
+      let leftSeparatorStyle;
+      let rightSeparatorStyle;
+      const separatorStyle = { height: customStyles.separatorStrokeWidth }
+      if(position === currentPosition) {
+        stepStyle = {
+          backgroundColor:customStyles.stepIndicatorCurrentColor,
+          borderWidth:customStyles.currentStepStrokeWidth,
+          borderColor:customStyles.stepIndicatorFinishedColor,
+          height:customStyles.currentStepIndicatorSize,
+          width:customStyles.currentStepIndicatorSize,
+          borderRadius:(customStyles.currentStepIndicatorSize) / 2
+        };
+        indicatorLabelStyle = { fontSize: customStyles.currentStepIndicatorLabelFontSize, color: customStyles.stepIndicatorLabelCurrentColor };
+        leftSeparatorStyle = { backgroundColor: customStyles.separatorFinishedColor };
+        rightSeparatorStyle = { backgroundColor: customStyles.separatorUnFinishedColor };
+      }
+      else if(position < currentPosition){
+        stepStyle = {
+          backgroundColor: customStyles.stepIndicatorFinishedColor,
+          height:customStyles.stepIndicatorSize,
+          width:customStyles.stepIndicatorSize,
+          borderRadius:(customStyles.stepIndicatorSize) / 2
+         };
+        indicatorLabelStyle = { fontSize: customStyles.stepIndicatorLabelFontSize, color: customStyles.stepIndicatorLabelFinishedColor };;
+        leftSeparatorStyle = { backgroundColor: customStyles.separatorFinishedColor };
+        rightSeparatorStyle = { backgroundColor: customStyles.separatorFinishedColor };
+      }
+      else {
+        stepStyle = {
+          backgroundColor: customStyles.stepIndicatorUnFinishedColor,
+          height:customStyles.stepIndicatorSize,
+          width:customStyles.stepIndicatorSize,
+          borderRadius:(customStyles.stepIndicatorSize) / 2
+         };
+        indicatorLabelStyle = { fontSize: customStyles.stepIndicatorLabelFontSize, color: customStyles.stepIndicatorLabelUnFinishedColor };
+        leftSeparatorStyle = { backgroundColor: customStyles.separatorUnFinishedColor };
+        rightSeparatorStyle = { backgroundColor: customStyles.separatorUnFinishedColor };
+      }
+
+
+      return [
+        <View key={'left-separator'} style={[separatorStyle,(position !== 0) ? leftSeparatorStyle : {backgroundColor:'transparent'},{width:this.getSeparatorWidth(), marginRight:-1*(customStyles.stepIndicatorSize/2)}]}/>,
+        <View key={'step-indicator'} style={[defaultStyle.step , stepStyle ]}>
+          <Text style={indicatorLabelStyle}>{ position + 1 }</Text>
+        </View>,
+        <View key={'right-separator'} style={[separatorStyle,(position !== stepCount - 1) ? rightSeparatorStyle : {backgroundColor:'transparent'},{width:this.getSeparatorWidth(), marginLeft:-1*(customStyles.stepIndicatorSize/2)}]}/>
+      ];
+    }
+
+    getSeparatorWidth = () => {
+      const { stepCount } = this.props;
+      const separatorWidth = (this.state.width/(stepCount*2));
+      return separatorWidth;
+    }
+
+    setCurrentPosition = () => {
+
+    }
   }
 
-  renderStrip = (position) => {
-    const { currentPosition } = this.props;
-    return(
-      <View style={[styles.separator , {width:this.getSeparatorWidth()}, position < currentPosition ? styles.separatorFinished : styles.separatorUnFinished]}/>
-    )
-  }
+  const defaultStyle =  StyleSheet.create({
+    container: {
+      backgroundColor:'transparent'
+    },
+    stepIndicatorContainer: {
+      flexDirection:'row',
+      alignItems:'center',
+      justifyContent:'space-around'
+    },
+    stepLabelsContainer: {
+      flexDirection:'row',
+      alignItems:'center',
+      justifyContent:'space-around',
+      paddingTop: 8
+    },
+    step: {
+      alignItems:'center',
+      justifyContent:'center',
+      zIndex: 2,
+      elevation:3
+    },
+    stepContainer: {
+      flex:1,
+      flexDirection:'row',
+      alignItems:'center',
+      justifyContent:'center'
+    },
+    stepLabel: {
+      flex:1,
+      fontSize:12,
+      textAlign:'center',
+      alignSelf:'flex-start'
+    }
+  });
 
-  getSeparatorWidth = () => {
-    const separatorWidth = (this.state.width - STEP_COUNT*STEP_INDICATOR_SIZE - (SELECTED_STEP_INDICATOR_SIZE_DELTA*2) + 32*(STEP_COUNT-1)) / (STEP_COUNT-1);
-    return separatorWidth;
-  }
+  StepIndicator.propTypes = {
+    currentPosition: PropTypes.number,
+    stepCount: PropTypes.number,
+    customStyles: PropTypes.object
+  };
 
-  setCurrentPosition = () => {
-
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-  },
-  stepIndicatorContainer: {
-    flexDirection:'row',
-    alignItems:'center',
-    justifyContent:'center'
-  },
-  step: {
-    height:STEP_INDICATOR_SIZE,
-    width:STEP_INDICATOR_SIZE,
-    borderRadius:STEP_INDICATOR_SIZE / 2,
-    alignItems:'center',
-    justifyContent:'center',
-    zIndex: 2
-  },
-  stepCurrent: {
-    backgroundColor:STEP_INDICATOR_COLOR_CURRENT,
-    borderWidth:5,
-    borderColor:STEP_INDICATOR_COLOR_FINISHED,
-    height:STEP_INDICATOR_SIZE+SELECTED_STEP_INDICATOR_SIZE_DELTA,
-    width:STEP_INDICATOR_SIZE+SELECTED_STEP_INDICATOR_SIZE_DELTA,
-    borderRadius:(STEP_INDICATOR_SIZE+SELECTED_STEP_INDICATOR_SIZE_DELTA) / 2,
-  },
-  stepFinished: {
-    backgroundColor:STEP_INDICATOR_COLOR_FINISHED
-  },
-  stepUnFinished: {
-    backgroundColor:STEP_INDICATOR_COLOR_UNFINISHED
-  },
-  separator: {
-    height:SEPARATOR_STROKE_WIDTH,
-    marginHorizontal: -16,
-  },
-  separatorFinished: {
-    backgroundColor:SEPARATOR_COLOR_FINISHED
-  },
-  separatorUnFinished: {
-    backgroundColor:SEPARATOR_COLOR_UNFINISHED
-  },
-  stepLabel: {
-    fontSize: STEP_INDICATOR_LABEL_SIZE,
-    color:'#ffffff'
-  },
-  stepLabelCurrent: {
-    color:STEP_INDICATOR_LABEL_COLOR_CURRENT,
-  },
-  stepLabelFinished: {
-    color:STEP_INDICATOR_LABEL_COLOR_FINISHED
-  },
-  stepLabelUnFinished: {
-    color:STEP_INDICATOR_LABEL_COLOR_UNFINISHED
-  },
-  stepContainer: {
-    flexDirection:'row',
-    alignItems:'center',
-    justifyContent:'center'
-  }
-});
-
-StepIndicator.propTypes = {
-  currentPosition: PropTypes.number
-};
-
-StepIndicator.defaultProps = {
-  currentPosition: 0
-}
+  StepIndicator.defaultProps = {
+    currentPosition: 0,
+    stepCount: 5,
+    customStyles: {}
+  };
