+++ example/.babelrc
