+++ example/.gitattributes
