+++ example/.watchmanconfig
