+++ example/App.js
@@ -1,14 +1,7 @@
-/**
-* Sample React Native App
-* https://github.com/facebook/react-native
-* @flow
-*/
-
 import React, { Component } from 'react';
 import { AppRegistry,StyleSheet,View, Text } from 'react-native';
 import ViewPager from 'react-native-viewpager';
-import StepIndicator from './StepIndicator';
-import NewStepIndicator from './NewStepIndicator';
+import StepIndicator from '../StepIndicator';
 const PAGES = ['Page 1','Page 2','Page 3','Page 4','Page 5'];
 
 export default class App extends Component {
@@ -29,8 +22,7 @@ export default class App extends Component {
     console.log("STATE PAGE " , this.state.currentPage);
     return (
       <View style={styles.container}>
-        <StepIndicator currentPosition={this.state.currentPage} />
-        <NewStepIndicator currentPosition={this.state.currentPage} labels={["label 1","label 2","label 3","label 4","label 5"]} />
+        <StepIndicator currentPosition={this.state.currentPage} labels={["label 1","label 2","label 3","label 4","label 5"]} />
         <ViewPager
           dataSource={this.state.dataSource}
           renderPage={this.renderViewPagerPage}
@@ -59,5 +51,3 @@ const styles = StyleSheet.create({
     alignItems:'center'
   }
 });
-
-AppRegistry.registerComponent('StepIndicator', () => App);
