+++ example/android/app/build.gradle
