+++ example/android/app/proguard-rules.pro
