+++ example/android/app/src/main/java/com/stepindicator/MainApplication.java
