+++ example/android/app/src/main/res/values/strings.xml
