+++ example/android/gradle.properties
