+++ example/android/keystores/debug.keystore.properties
