+++ example/android/settings.gradle
