+++ example/ios/StepIndicator.xcodeproj/xcshareddata/xcschemes/StepIndicator.xcscheme
