+++ example/ios/StepIndicator/AppDelegate.h
