+++ example/ios/StepIndicator/AppDelegate.m
