+++ example/ios/StepIndicator/Base.lproj/LaunchScreen.xib
