+++ example/ios/StepIndicator/Images.xcassets/AppIcon.appiconset/Contents.json
