+++ example/ios/StepIndicator/Info.plist
