+++ example/ios/StepIndicatorTests/Info.plist
