+++ example/ios/StepIndicatorTests/StepIndicatorTests.m
