+++ package.json
@@ -1,23 +1,30 @@
 {
-  "name": "StepIndicator",
+  "name": "react-native-step-indicator",
   "version": "0.0.1",
-  "private": true,
+  "description": "A simple react-native implementation step indicator widget compatible with the ViewPager.",
+  "main": "StepIndicator.js",
+  "directories": {
+    "example": "example"
+  },
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "dependencies": {
-    "react": "15.4.2",
-    "react-native": "0.40.0",
-    "react-native-viewpager": "^0.2.13"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/24ark/react-native-step-indicator.git"
   },
-  "devDependencies": {
-    "babel-jest": "18.0.0",
-    "babel-preset-react-native": "1.9.1",
-    "jest": "18.1.0",
-    "react-test-renderer": "15.4.2"
+  "keywords": [
+    "react-component",
+    "ios",
+    "android",
+    "react-native",
+    "step-indicator",
+    "viewpager"
+  ],
+  "author": "Arkit Vora",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/24ark/react-native-step-indicator/issues"
   },
-  "jest": {
-    "preset": "react-native"
-  }
+  "homepage": "https://github.com/24ark/react-native-step-indicator#readme"
 }
